withSecurityPageIsEnabled.js -> ../withSecurityPageIsEnabled.js
